+++ fe/fe-core/src/test/java/com/starrocks/pseudocluster/BeTxnManager.java
@@ -1,7 +1,6 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
 package com.starrocks.pseudocluster;
 
-import com.starrocks.common.AlreadyExistsException;
 import com.starrocks.common.UserException;
 import com.starrocks.thrift.TFinishTaskRequest;
 import com.starrocks.thrift.TPartitionVersionInfo;
@@ -48,7 +47,8 @@ public class BeTxnManager {
         Map<Long, TxnTabletInfo> tablets = tinfo.partitions.computeIfAbsent(partitionId, k -> new HashMap<>());
         TxnTabletInfo tabletInfo = tablets.get(tablet.id);
         if (tabletInfo != null) {
-            throw new AlreadyExistsException("txn:" + txnId + " tablet:" + tablet.id + " already exists");
+            LOG.info("txn:" + txnId + " tablet:" + tablet.id + " already exists, multi node sink");
+            return;
         }
         tabletInfo = tablets.computeIfAbsent(tablet.id, id -> new TxnTabletInfo(id));
         tabletInfo.rowset = rowset;
