+++ fe/fe-core/src/test/java/com/starrocks/transaction/ConcurrentTxnNewPublishTest.java
@@ -9,8 +9,6 @@ import java.sql.SQLException;
 public class ConcurrentTxnNewPublishTest extends ConcurrentTxnTest {
     @Override
     void setup() throws SQLException {
-        runTime = 20;
-        withRead = true;
         Config.enable_new_publish_mechanism = false;
     }
 }
